+++ src/backend/resolvers/mutations/businessObjectModal/businessObjectModal/index.js
@@ -1,0 +1,9 @@
+const createBusinessObjectModal = require('./create')
+// const updateWorkbook = require('./update')
+// const deleteWorkbook = require('./delete')
+
+module.exports = {
+  createBusinessObjectModal,
+  // updateWorkbook,
+  // deleteWorkbook,
+}