+++ src/backend/resolvers/mutations/businessObjectModal/field/index.js
@@ -1,0 +1,5 @@
+const createBusinessObjectModalField = require('./create')
+
+module.exports = {
+  createBusinessObjectModalField,
+}