+++ src/backend/resolvers/mutations/businessObjectModal/index.js
@@ -1,0 +1,11 @@
+const businessObjectModalResolvers = require('./businessObjectModal')
+const tagResolvers = require('./tag')
+const sectionResolvers = require('./section')
+const fieldResolvers = require('./field')
+
+module.exports = {
+  ...businessObjectModalResolvers,
+  ...tagResolvers,
+  ...sectionResolvers,
+  ...fieldResolvers,
+}