+++ src/backend/resolvers/mutations/businessObjectModal/section/index.js
@@ -1,0 +1,5 @@
+const createBusinessObjectModalSection = require('./create')
+
+module.exports = {
+  createBusinessObjectModalSection,
+}