+++ src/backend/resolvers/mutations/businessObjectModal/tag/index.js
@@ -1,0 +1,5 @@
+const createBusinessObjectModalTag = require('./create')
+
+module.exports = {
+  createBusinessObjectModalTag,
+}