+++ src/backend/typeDefs/mutations/businessObjectModal/businessObjectModal/index.js
@@ -1,0 +1,5 @@
+const createBusinessObjectModalTypeDefs = require('./create')
+
+module.exports = [
+  createBusinessObjectModalTypeDefs,
+]