+++ src/backend/typeDefs/mutations/businessObjectModal/field/index.js
@@ -1,0 +1,5 @@
+const createBusinessObjectModalFieldTypeDefs = require('./create')
+
+module.exports = [
+  createBusinessObjectModalFieldTypeDefs,
+]