+++ src/backend/typeDefs/mutations/businessObjectModal/section/create.js
@@ -1,0 +1,11 @@
+const { gql } = require('apollo-server-express')
+
+const createBusinessObjectModalSectionTypeDefs = gql`
+  input CreateBusinessObjectModalSectionInput {
+    label: String!
+    modalId: ID!
+    tagId: ID!
+  }
+`
+
+module.exports = createBusinessObjectModalSectionTypeDefs