+++ src/backend/typeDefs/mutations/businessObjectModal/section/index.js
@@ -1,0 +1,5 @@
+const createBusinessObjectModalSectionTypeDefs = require('./create')
+
+module.exports = [
+  createBusinessObjectModalSectionTypeDefs,
+]