+++ src/backend/typeDefs/mutations/businessObjectModal/tag/index.js
@@ -1,0 +1,5 @@
+const createBusinessObjectModalTagTypeDefs = require('./create')
+
+module.exports = [
+  createBusinessObjectModalTagTypeDefs,
+]