+++ src/frontend/api/mutations/bomConfigs.js
@@ -1,0 +1,69 @@
+import gql from 'graphql-tag'
+
+export const CREATE_BOM_CONFIG_FIELD = gql`
+  mutation CreateBusinessObjectModalField($input: CreateBusinessObjectModalFieldInput!) {
+    createBusinessObjectModalField(input: $input) {
+      _id
+      label
+      inputComponent
+      inputProps
+    }
+  }
+`
+
+export const CREATE_BOM_CONFIG_SECTION = gql`
+  mutation CreateBusinessObjectModalSection($input: CreateBusinessObjectModalSectionInput!) {
+    createBusinessObjectModalSection(input: $input) {
+      _id
+      label
+      fields {
+        _id
+        label
+        inputComponent
+        inputProps
+      }
+    }
+  }
+`
+
+export const CREATE_BOM_CONFIG_TAB = gql`
+  mutation CreateBusinessObjectModalTab($input: CreateBusinessObjectModalTagInput!) {
+    createBusinessObjectModalTag(input: $input) {
+      _id
+      label
+      sections {
+        _id
+        label
+        fields {
+          _id
+          label
+          inputComponent
+          inputProps
+        }
+      }
+    }
+  }
+`
+
+export const CREATE_BOM_CONFIG = gql`
+  mutation CreateBusinessObjectModal($input: CreateBusinessObjectModalInput!) {
+    createBusinessObjectModal(input: $input) {
+      _id
+      label
+      tags {
+        _id
+        label
+        sections {
+          _id
+          label
+          fields {
+            _id
+            label
+            inputComponent
+            inputProps
+          }
+        }
+      }
+    }
+  }
+`